+++ src/llama_cloud/types/extraction/__init__.py
@@ -10,11 +10,9 @@ from .run_get_params import RunGetParams as RunGetParams
 from .job_file_params import JobFileParams as JobFileParams
 from .job_list_params import JobListParams as JobListParams
 from .run_list_params import RunListParams as RunListParams
-from .job_batch_params import JobBatchParams as JobBatchParams
 from .job_create_params import JobCreateParams as JobCreateParams
 from .job_list_response import JobListResponse as JobListResponse
 from .run_delete_params import RunDeleteParams as RunDeleteParams
-from .job_batch_response import JobBatchResponse as JobBatchResponse
 from .extract_config_param import ExtractConfigParam as ExtractConfigParam
 from .job_get_result_params import JobGetResultParams as JobGetResultParams
 from .run_get_by_job_params import RunGetByJobParams as RunGetByJobParams